+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - 'jot' now supports relative date parsing, such as setting a due date field to 'tomorrow' or 'next Tuesday'
 - Home-rooted paths, like "~/Downloads/example.txt" in arguments are now expanded when processed.
 - Configuration of storage providers is now made via appsettings.json.
+- 'jot' now supports a --filter option for the `things` command for complex filtering expressions.
 
 ### Modified